rename factory classes/interfaces and eliminate unneeded methods

+++ kas-mq-admin/src/main/java/com/kas/mq/console/cmds/delete/DeleteQueueCommand.java
@@ -1,6 +1,7 @@
 package com.kas.mq.console.cmds.delete;
 
 import com.kas.infra.utils.ConsoleUtils;
+import com.kas.infra.utils.Validators;
 import com.kas.mq.console.ACommand;
 import com.kas.mq.internal.MqContextConnection;
 
@@ -42,6 +43,9 @@ public class DeleteQueueCommand extends ACommand
    */
   public void exec(MqContextConnection conn)
   {
+    if (!Validators.isQueueName(mName))
+      throw new IllegalArgumentException("NAME was not specified or invalid: [" + mName + ']');
+    
     conn.deleteQueue(mName, mForce);
     ConsoleUtils.writeln("%s", conn.getResponse());
   }
